Move files to prepare Moved some files to prepare for the next step in moving to slash commands. Only callable commands can be in the /commands folder because fs.readDir will be used to get all commands in help, command registration, and command execution.
