+++ events/ready.mjs
@@ -1,8 +1,8 @@
 import guildCreate  from "./guildCreate.mjs";
-import reactRole    from "../commands/reactrole/collect.mjs";
-import pollCollect  from "../commands/poll/collect.mjs";
-import unmute       from "../commands/mute/unmute.mjs";
-import reactRoleOne from "../commands/reactroleone/collect.mjs";
+import reactRole    from "../utils/reactrole.mjs";
+import pollCollect  from "../utils/poll.mjs";
+import unmute       from "../utils/unmute.mjs";
+import reactRoleOne from "../utils/reactroleone.mjs";
 import muterole from "../utils/mutrole.mjs";
 import clearMjs from "../commands/clear/clear.mjs";
 import discord from "discord.js";
