+++ utils/funnytext/blarb.mjs
