+++ utils/funnytext/lego.mjs
