+++ utils/funnytext/piglatin.mjs
