+++ utils/funnytext/reverse.mjs
