+++ utils/funnytext/shakespeare.mjs
