+++ utils/funnytext/uwu.mjs
